WIP: factor out Fps class
