+++ js/index.js
@@ -1,48 +1,9 @@
 // import { memory } from "../rust/pkg/rust_wasm_gol_bg.wasm";
+import { Fps } from "./fps.js";
 import { vertex, fragment } from "./shaders.js";
 import { Program, Buffer, VertexArrayObject, clear, drawArrays, FragmentShader, VertexShader } from "./bettergl.js";
 
-const fps = new class {
-  constructor() {
-    this.fps = document.getElementById("fps");
-    this.frames = [];
-    this.lastFrameTimeStamp = performance.now();
-  }
-
-  render() {
-    const now = performance.now();
-    const delta = now - this.lastFrameTimeStamp;
-    this.lastFrameTimeStamp = now;
-    const fps =
-      // delta: milliseconds
-      // 1 / (delta / 1000)
-      // =
-      1 / delta * 1000;
-
-    this.frames.push(fps);
-    if (this.frames.length > 100) {
-      this.frames.shift();
-    }
-
-    let min = Infinity;
-    let max = -Infinity;
-    let sum = 0;
-    for (let i = 0; i < this.frames.length; i++) {
-      sum += this.frames[i];
-      min = Math.min(this.frames[i], min);
-      max = Math.max(this.frames[i], max);
-    }
-    let mean = sum / this.frames.length;
-
-    this.fps.textContent = `
-Frames per Second:
-         latest = ${Math.round(fps)}
-avg of last 100 = ${Math.round(mean)}
-min of last 100 = ${Math.round(min)}
-max of last 100 = ${Math.round(max)}
-`.trim();
-  }
-};
+const fps = new Fps();
 
 import("../rust/pkg/rust_wasm_gol.js").then(module => {
   const { start, Cell, World } = module;
